Reserve Billing Credits on Workflow Execution
This commit moves the billing service client into the metering entity and makes a call to reserve credits when
the metering reserve call is make from the engine. This reduces touch points within the engine such that it makes
a single reserve call to metering and credits are automatically handled.

Currently the credit type is inferred from the capability type. This is hard-coded for now for demo purposes only.