feat: add token deployment implementation for EVM and Solana (#17847)
This commit introduces the implementation for deploying tokens on EVM
and Solana networks using the Operations API. These changesets will
eventually replace the existing token deployment logic in the `common`
package.

By introducing a new `tokens` package, we aim to provide more structure
to how changesets related to tokens are handled, and this also serves
as an example of how to use ChangeSetV2 and Operations API for building
Changesets.