Metering steps not included in metering report when in metering mode (#19096)
* Metering steps not included in metering report when in metering mode

A set of cases came up for incomplete metering report data specifically for when a billing client was missing. Some
of these cases included messages of Settle being called before Deduct. The solution was to ensure the `Report.ready`
property was set in cases where `Reserve` was called with or without errors.

The ready state was being set in when metering mode was turned on, but since some metering mode cases were introduced
before `Reserve` _could_ be called, it made more sense to set the ready state for all cases of `Reserve` and change
the property name for more clarity.

Additionally, all cases where metering mode could be turned on use the helper function. The helper function was updated
to log metering mode details only once in an effort to not clutter the logs, but to capture all encountered errors in
the metering report.

* relax the failure handling in metering:

* add log to verify report format

* log debug step ref info

* include step refs in the same log message

* update tests

* deep equal comparison of metering steps

* add Set function to balance store to apply reserve amount

* adding testing

* fix linting

---------

Co-authored-by: Patrick Huie <patrick.huie@smartcontract.com>